backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP requests (if needed for external APIs)
httpx==0.25.1
requests==2.31.0

# Wikipedia & PDF extraction
wikipedia==1.4.0
langchain-community==0.0.38
pypdf==3.17.4

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

bcrypt==3.2.2
groq
google-genai